@@ -1,35 +1,34 @@
 name: mono_kit
 description: A collection of convenient widgets and utils made by mono. Under construction now.
-version: 0.18.2-dev
+version: 0.18.3
 homepage: https://github.com/mono0926/flutter_mono_kit
 environment:
-  sdk: ">=2.6.0 <3.0.0"
+  sdk: ">=2.12.0-0 <3.0.0"
   flutter: ">=1.14.6 <2.0.0"
 dependencies:
   adaptive_dialog: ">=0.7.0 <1.0.0"
+  disposable_provider: ">=2.0.0 <3.0.0"
   flutter:
     sdk: flutter
-  flutter_cache_manager: ">=1.4.0 <2.0.0"
-  image_picker: ">=0.6.0 <1.0.0"
-  rxdart: ">=0.24.0 <1.0.0"
-  url_launcher: ">=5.4.0 <6.0.0"
-  disposable_provider: ">=2.0.0 <3.0.0"
   freezed_annotation: ">=0.11.0 <1.0.0"
-  nested: ">=0.0.4 <1.0.0"
-  hooks_riverpod: ">=0.9.0 <1.0.0"
+  image_picker: ">=0.6.0 <1.0.0"
+  rxdart: ">=0.26.0-0 <1.0.0"
+  url_launcher: ">=5.4.0 <6.0.0"
+  nested: ">=0.0.5-0 <1.0.0"
+  hooks_riverpod: ">=0.13.0-0 <1.0.0"
   quiver: ">=2.0.0 <3.0.0"
-  stream_transform: ">=1.0.0 <2.0.0"
-  subscription_holder: ">=2.0.0 <3.0.0"
+  stream_transform: ">=2.0.0 <3.0.0"
+  subscription_holder: ">=2.1.0 <3.0.0"
   simple_logger: ">=1.6.0 <2.0.0"
   tinycolor: ">=1.0.0 <2.0.0"
-  flutter_svg: ">=0.18.0 <1.0.0"
+  flutter_svg: ">=0.20.0-0 <1.0.0"
   state_notifier: ">=0.6.0 <1.0.0"
   flutter_hooks: ">=0.13.0 <1.0.0"
 dev_dependencies:
   build_runner:
   flutter_test:
     sdk: flutter
-  freezed:
+  freezed: ">=0.13.0-- <0.14.0"
   pedantic_mono:
 flutter:
   plugin:
