@@ -70,15 +70,15 @@ class __BodyState extends State<_Body> {
           Row(
             mainAxisAlignment: MainAxisAlignment.spaceEvenly,
             children: <Widget>[
-              RaisedButton(
+              ElevatedButton(
                 child: const Text('TOP'),
                 onPressed: _scrollController.scrollToTop,
               ),
-              RaisedButton(
+              ElevatedButton(
                 child: const Text('UP️'),
                 onPressed: _scrollController.scrollToPreviousPage,
               ),
-              RaisedButton(
+              ElevatedButton(
                 child: const Text('↑'),
                 onPressed: _scrollController.scrollUp,
               ),
@@ -87,15 +87,15 @@ class __BodyState extends State<_Body> {
           Row(
             mainAxisAlignment: MainAxisAlignment.spaceEvenly,
             children: <Widget>[
-              RaisedButton(
+              ElevatedButton(
                 child: const Text('BOTTOM️'),
                 onPressed: _scrollController.scrollToBottom,
               ),
-              RaisedButton(
+              ElevatedButton(
                 child: const Text('DOWN'),
                 onPressed: _scrollController.scrollToNextPage,
               ),
-              RaisedButton(
+              ElevatedButton(
                 child: const Text('↓'),
                 onPressed: _scrollController.scrollDown,
               ),
