@@ -4,8 +4,8 @@ import 'modal_page_route.dart';
 
 class FadePageRoute<T> extends ModalPageRoute<T> {
   FadePageRoute({
-    @required WidgetBuilder builder,
-    RouteSettings settings,
+    required WidgetBuilder builder,
+    RouteSettings? settings,
   }) : super(
           builder: builder,
           settings: settings,
