@@ -2,12 +2,12 @@ import 'package:flutter/material.dart';
 
 class AnimatedExpansionVisibility extends ImplicitlyAnimatedWidget {
   const AnimatedExpansionVisibility({
-    Key key,
-    @required this.child,
-    @required this.isVisible,
+    Key? key,
+    required this.child,
+    required this.isVisible,
     this.alignment,
-    Duration duration,
-    Curve curve,
+    Duration? duration,
+    Curve? curve,
   }) : super(
           key: key,
           duration: duration ?? const Duration(milliseconds: 200),
@@ -16,7 +16,7 @@ class AnimatedExpansionVisibility extends ImplicitlyAnimatedWidget {
 
   final Widget child;
   final bool isVisible;
-  final Alignment alignment;
+  final Alignment? alignment;
 
   @override
   _AnimatedExpansionVisibilityState createState() =>
@@ -25,7 +25,7 @@ class AnimatedExpansionVisibility extends ImplicitlyAnimatedWidget {
 
 class _AnimatedExpansionVisibilityState
     extends AnimatedWidgetBaseState<AnimatedExpansionVisibility> {
-  Tween<double> _heightFactorTween;
+  late Tween<double> _heightFactorTween;
 
   @override
   Widget build(BuildContext context) {
@@ -48,6 +48,6 @@ class _AnimatedExpansionVisibilityState
       _heightFactorTween,
       widget.isVisible ? 1.0 : 0.0,
       (dynamic heightFactor) => Tween<double>(begin: heightFactor as double),
-    ) as Tween<double>;
+    )! as Tween<double>;
   }
 }
