@@ -4,9 +4,9 @@ import 'package:url_launcher/url_launcher.dart';
 
 class LinkTextSpan extends TextSpan {
   LinkTextSpan({
-    @required TextStyle style,
-    @required String url,
-    String text,
+    required TextStyle style,
+    required String url,
+    String? text,
   }) : super(
           style: style,
           text: text ?? url,
