@@ -3,20 +3,17 @@ import 'package:mono_kit/mono_kit.dart';
 
 class LoadingSwitcher extends StatelessWidget {
   const LoadingSwitcher({
-    Key key,
+    Key? key,
     this.child,
     this.timedOutChild = const CircularProgressIndicator(),
     this.initialChild = const SizedBox(),
     this.timeout = const Duration(milliseconds: 200),
     this.switchDuration = const Duration(milliseconds: 100),
-  })  : assert(timedOutChild != null),
-        assert(initialChild != null),
-        assert(timeout != null),
-        super(key: key);
+  }) : super(key: key);
 
   final Widget timedOutChild;
   final Widget initialChild;
-  final Widget child;
+  final Widget? child;
   final Duration timeout;
   final Duration switchDuration;
 
