@@ -8,9 +8,9 @@ typedef OverflowDetector = void Function(double overflowHeight);
 class OverflowDetectableText extends StatelessWidget {
   const OverflowDetectableText(
     this.data, {
-    Key key,
-    @required this.style,
-    @required this.detector,
+    Key? key,
+    required this.style,
+    required this.detector,
     this.overflow = TextOverflow.clip,
   }) : super(key: key);
 
@@ -38,7 +38,7 @@ class OverflowDetectableText extends StatelessWidget {
           0.0, // ignore: prefer_int_literals
           painter.height - constraints.maxHeight,
         );
-        SchedulerBinding.instance
+        SchedulerBinding.instance!
             .addPostFrameCallback((_) => detector(overflowHeight));
         return Text.rich(
           span,
