@@ -3,25 +3,25 @@ import 'package:nested/nested.dart';
 
 class TilePadding extends SingleChildStatelessWidget {
   const TilePadding({
-    Key key,
-    Widget child,
+    Key? key,
+    Widget? child,
     this.horizontal,
   }) : super(
           key: key,
           child: child,
         );
 
-  final double horizontal;
+  final double? horizontal;
 
   @override
-  Widget buildWithChild(BuildContext context, Widget child) {
+  Widget buildWithChild(BuildContext context, Widget? child) {
     return SafeArea(
       top: false,
       bottom: false,
       minimum: EdgeInsets.symmetric(
         horizontal: horizontal ?? 16,
       ),
-      child: child,
+      child: child!,
     );
   }
 }
