@@ -7,7 +7,7 @@ import 'plugins.dart';
 class IosInfoPlistReader {
   static const _platform = MethodChannel(kMethodChannelName);
 
-  Future<String> readString(String key) async {
+  Future<String?> readString(String key) async {
     assert(Platform.isIOS);
     return _platform.invokeMethod(
       'info_plist/read_string',
