@@ -7,8 +7,8 @@ void main() {
   group(
     'changedValuePair(Stream)',
     () {
-      PublishSubject<String> _subject;
-      Stream<ChangedValuePair<String>> _target;
+      late PublishSubject<String?> _subject;
+      late Stream<ChangedValuePair<String?>> _target;
       setUp(() {
         _subject = PublishSubject();
         _target = _subject.changedValuePair();
@@ -40,8 +40,8 @@ void main() {
   group(
     'changedValuePair(ValueStream(initialValue: null))',
     () {
-      BehaviorSubject<String> _subject;
-      Stream<ChangedValuePair<String>> _target;
+      late BehaviorSubject<String?> _subject;
+      late Stream<ChangedValuePair<String?>> _target;
       setUp(() {
         _subject = BehaviorSubject();
         _target = _subject.changedValuePair();
@@ -58,13 +58,13 @@ void main() {
             expect(pair.current, isNot(equals(pair.previous)));
           });
           _subject
-            ..value = 'a'
-            ..value = 'a'
-            ..value = 'b'
-            ..value = 'b'
-            ..value = 'c'
-            ..value = null
-            ..value = 'a';
+            ..add('a')
+            ..add('a')
+            ..add('b')
+            ..add('b')
+            ..add('c')
+            ..add(null)
+            ..add('a');
         },
       );
     },
@@ -73,8 +73,8 @@ void main() {
   group(
     'changedValuePair(ValueStream(initialValue: a))',
     () {
-      BehaviorSubject<String> _subject;
-      Stream<ChangedValuePair<String>> _target;
+      late BehaviorSubject<String?> _subject;
+      late Stream<ChangedValuePair<String?>> _target;
       setUp(() {
         _subject = BehaviorSubject.seeded('a');
         _target = _subject.changedValuePair();
@@ -91,13 +91,13 @@ void main() {
             expect(pair.current, isNot(equals(pair.previous)));
           });
           _subject
-            ..value = 'a'
-            ..value = 'a'
-            ..value = 'b'
-            ..value = 'b'
-            ..value = 'c'
-            ..value = null
-            ..value = 'a';
+            ..add('a')
+            ..add('a')
+            ..add('b')
+            ..add('b')
+            ..add('c')
+            ..add(null)
+            ..add('a');
           await expectLater(
             await _target.first,
             equals(const ChangedValuePair(previous: 'a', current: 'a')),
